# Conflicts:
#	frontend-admin-site/public_html/i18n/app.en.json
#	frontend-admin-site/public_html/js/app.js

+++ backend-rest-api/src/main/java/com/nitsoft/ecommerce/database/model/UserAddress.java
@@ -36,7 +36,7 @@ public class UserAddress implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "adress_id")
-    private Integer adressId;
+    private Long adressId;
     
     @Basic(optional = false)
     @Column(name = "user_id")
